move to internal/license instead
