+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/license_check_handler_test.go
@@ -21,7 +21,7 @@ import (
 
 func TestNewLicenseCheckHandler(t *testing.T) {
 	makeToken := func(licenseKey string) []byte {
-		token := licensing.GenerateLicenseKeyBasedAccessToken(licenseKey)
+		token := license.GenerateLicenseKeyBasedAccessToken(licenseKey)
 		return []byte(token)
 	}
 	now := time.Now()
