+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/licenses_db.go
@@ -9,11 +9,11 @@ import (
 	"github.com/keegancsmith/sqlf"
 	"github.com/lib/pq"
 
-	"github.com/sourcegraph/sourcegraph/internal/licensing"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/hashutil"
 	"github.com/sourcegraph/sourcegraph/internal/license"
+	"github.com/sourcegraph/sourcegraph/internal/licensing"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -114,7 +114,7 @@ func (s dbLicenses) GetByAccessToken(ctx context.Context, accessToken string) (*
 		return mocks.licenses.GetByToken(accessToken)
 	}
 
-	contents, err := licensing.ExtractLicenseKeyBasedAccessTokenContents(accessToken)
+	contents, err := license.GenerateLicenseKeyBasedAccessToken(accessToken)
 	if err != nil {
 		return nil, errTokenInvalid
 	}
