+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/licenses_db_test.go
@@ -12,10 +12,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/sourcegraph/sourcegraph/internal/licensing"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
 	"github.com/sourcegraph/sourcegraph/internal/license"
+	"github.com/sourcegraph/sourcegraph/internal/licensing"
 	"github.com/sourcegraph/sourcegraph/internal/timeutil"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/lib/pointers"
@@ -123,7 +123,7 @@ func TestGetByToken(t *testing.T) {
 	}{
 		{
 			name:  "ok",
-			token: licensing.GenerateLicenseKeyBasedAccessToken("key"),
+			token: license.GenerateLicenseKeyBasedAccessToken("key"),
 			want:  &license.ID,
 		},
 		{
