+++ internal/license/BUILD.bazel
@@ -3,10 +3,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "license",
-    srcs = ["license.go"],
+    srcs = [
+        "access_token.go",
+        "license.go",
+    ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/license",
     visibility = ["//:__subpackages__"],
     deps = [
+        "//internal/hashutil",
         "//lib/errors",
         "@org_golang_x_crypto//ssh",
     ],
@@ -15,7 +19,15 @@ go_library(
 go_test(
     name = "license_test",
     timeout = "short",
-    srcs = ["license_test.go"],
+    srcs = [
+        "access_token_test.go",
+        "license_test.go",
+    ],
     embed = [":license"],
-    deps = ["@org_golang_x_crypto//ssh"],
+    deps = [
+        "@com_github_hexops_autogold_v2//:autogold",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+        "@org_golang_x_crypto//ssh",
+    ],
 )
