+++ internal/license/access_token.go
@@ -1,4 +1,4 @@
-package licensing
+package license
 
 import (
 	"encoding/hex"
