+++ internal/licensing/BUILD.bazel
@@ -4,7 +4,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "licensing",
     srcs = [
-        "access_token.go",
         "check.go",
         "codygateway.go",
         "conf.go",
@@ -25,7 +24,6 @@ go_library(
         "//internal/env",
         "//internal/errcode",
         "//internal/goroutine",
-        "//internal/hashutil",
         "//internal/httpcli",
         "//internal/license",
         "//internal/redispool",
@@ -42,7 +40,6 @@ go_test(
     name = "licensing_test",
     timeout = "short",
     srcs = [
-        "access_token_test.go",
         "check_test.go",
         "codygateway_test.go",
         "features_test.go",
@@ -61,7 +58,6 @@ go_test(
         "@com_github_derision_test_glock//:glock",
         "@com_github_gomodule_redigo//redis",
         "@com_github_google_go_cmp//cmp",
-        "@com_github_hexops_autogold_v2//:autogold",
         "@com_github_sourcegraph_log//logtest",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
