+++ internal/licensing/check.go
@@ -15,6 +15,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 	"github.com/sourcegraph/sourcegraph/internal/httpcli"
+	"github.com/sourcegraph/sourcegraph/internal/license"
 	"github.com/sourcegraph/sourcegraph/internal/redispool"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
@@ -173,7 +174,7 @@ func StartLicenseCheck(originalCtx context.Context, logger log.Logger, siteID st
 		ctxWithCancel, cancel = context.WithCancel(originalCtx)
 
 		prevLicenseToken, _ := store.Get(prevLicenseTokenKey).String()
-		licenseToken := GenerateLicenseKeyBasedAccessToken(conf.Get().LicenseKey)
+		licenseToken := license.GenerateLicenseKeyBasedAccessToken(conf.Get().LicenseKey)
 		var initialWaitInterval time.Duration = 0
 		if prevLicenseToken == licenseToken {
 			initialWaitInterval, _ = calcDurationSinceLastCalled(glock.NewRealClock())
